Move the shared web view code out of the paywall component folder
WebViewOrigin and the navigation-response status check are not specific to the
web_view component, and already had a caller outside Components in
Cache/WebBundlePrewarmer. Put them in RevenueCatUI/WebView/ so a second web view
can use them without reaching into the paywall component tree.

WebViewNavigationPolicy becomes WebViewComponentNavigationPolicy, since it is
now one of two policies and governs only the component.

Moves and a rename, no behaviour change.
