+++ RevenueCatUI/WebView/WebViewOrigin.swift
@@ -2,7 +2,7 @@
 //  Copyright RevenueCat Inc. All Rights Reserved.
 //
 
-#if !os(tvOS) && canImport(WebKit) // For Paywalls V2
+#if !os(tvOS) && canImport(WebKit)
 
 import Foundation
 import WebKit
@@ -81,34 +81,6 @@ struct WebViewOrigin: Equatable {
 
     private static func isDefaultPort(_ port: Int, scheme: String) -> Bool {
         (scheme == "https" && port == 443) || (scheme == "http" && port == 80)
-    }
-
-}
-
-@available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
-enum WebViewNavigationPolicy {
-
-    // Non-https navigation is blocked on any frame. Cross-origin navigation is additionally blocked
-    // on the main frame (same-origin different-path navigation stays allowed), which makes
-    // cross-origin message races structurally impossible. Cross-origin sub-frame loads are not
-    // blocked here; isolation for those is left to the server-provided CSP (`frame-src` falls back
-    // to `default-src 'self'`).
-    static func policy(for url: URL?, isMainFrame: Bool, expectedOrigin: WebViewOrigin) -> WKNavigationActionPolicy {
-        guard let origin = WebViewOrigin(url: url), origin.isHTTPS else {
-            return .cancel
-        }
-        guard isMainFrame else {
-            return .allow
-        }
-        return origin == expectedOrigin ? .allow : .cancel
-    }
-
-    // WebKit treats an HTTP 4xx/5xx as a *successful* navigation (the error body renders and `didFail*`
-    // never fires), so a navigation response's status code is the only signal that the main document
-    // actually failed to load. Sub-frame/sub-resource errors are ignored so a single failing asset
-    // doesn't remove the whole component.
-    static func isTerminalHTTPError(statusCode: Int, isMainFrame: Bool) -> Bool {
-        return isMainFrame && statusCode >= 400
     }
 
 }
