+++ Tests/RevenueCatUITests/PaywallsV2/WebViewComponentNavigationPolicyTests.swift
@@ -9,13 +9,13 @@ import XCTest
 import WebKit
 
 @available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
-final class WebViewNavigationPolicyTests: TestCase {
+final class WebViewComponentNavigationPolicyTests: TestCase {
 
     private let expectedOrigin = WebViewOrigin(string: "https://example.com")!
 
     func testSameOriginMainFrameDifferentPathIsAllowed() {
         XCTAssertEqual(
-            WebViewNavigationPolicy.policy(
+            WebViewComponentNavigationPolicy.policy(
                 for: URL(string: "https://example.com/promo/step-two.html")!,
                 isMainFrame: true,
                 expectedOrigin: self.expectedOrigin
@@ -26,7 +26,7 @@ final class WebViewNavigationPolicyTests: TestCase {
 
     func testCrossOriginMainFrameIsCancelled() {
         XCTAssertEqual(
-            WebViewNavigationPolicy.policy(
+            WebViewComponentNavigationPolicy.policy(
                 for: URL(string: "https://evil.example/phish.html")!,
                 isMainFrame: true,
                 expectedOrigin: self.expectedOrigin
@@ -37,7 +37,7 @@ final class WebViewNavigationPolicyTests: TestCase {
 
     func testMainFramePortMismatchIsCancelled() {
         XCTAssertEqual(
-            WebViewNavigationPolicy.policy(
+            WebViewComponentNavigationPolicy.policy(
                 for: URL(string: "https://example.com:8443/next")!,
                 isMainFrame: true,
                 expectedOrigin: self.expectedOrigin
@@ -48,7 +48,7 @@ final class WebViewNavigationPolicyTests: TestCase {
 
     func testNonHttpsIsCancelledOnAnyFrame() {
         XCTAssertEqual(
-            WebViewNavigationPolicy.policy(
+            WebViewComponentNavigationPolicy.policy(
                 for: URL(string: "http://example.com/promo/index.html")!,
                 isMainFrame: false,
                 expectedOrigin: self.expectedOrigin
@@ -56,7 +56,7 @@ final class WebViewNavigationPolicyTests: TestCase {
             .cancel
         )
         XCTAssertEqual(
-            WebViewNavigationPolicy.policy(
+            WebViewComponentNavigationPolicy.policy(
                 for: URL(string: "custom://example.com/")!,
                 isMainFrame: true,
                 expectedOrigin: self.expectedOrigin
@@ -68,7 +68,7 @@ final class WebViewNavigationPolicyTests: TestCase {
     func testCrossOriginHttpsSubFrameIsAllowed() {
         // Sub-frame isolation is expected from the server-provided CSP, not this navigation policy.
         XCTAssertEqual(
-            WebViewNavigationPolicy.policy(
+            WebViewComponentNavigationPolicy.policy(
                 for: URL(string: "https://other.example.com/frame.html")!,
                 isMainFrame: false,
                 expectedOrigin: self.expectedOrigin
@@ -79,7 +79,7 @@ final class WebViewNavigationPolicyTests: TestCase {
 
     func testNilURLMainFrameIsCancelled() {
         XCTAssertEqual(
-            WebViewNavigationPolicy.policy(
+            WebViewComponentNavigationPolicy.policy(
                 for: nil,
                 isMainFrame: true,
                 expectedOrigin: self.expectedOrigin
@@ -91,7 +91,7 @@ final class WebViewNavigationPolicyTests: TestCase {
     func testNonCanonicalTargetCaseAndDefaultPortIsAllowed() {
         // The navigated URL is normalized before comparison, so case and explicit default port match.
         XCTAssertEqual(
-            WebViewNavigationPolicy.policy(
+            WebViewComponentNavigationPolicy.policy(
                 for: URL(string: "HTTPS://Example.COM:443/promo/index.html")!,
                 isMainFrame: true,
                 expectedOrigin: self.expectedOrigin
@@ -104,31 +104,13 @@ final class WebViewNavigationPolicyTests: TestCase {
         // The expected origin is normalized at construction, so a non-canonical spelling still matches.
         let expectedOrigin = WebViewOrigin(string: "https://Example.COM:443")!
         XCTAssertEqual(
-            WebViewNavigationPolicy.policy(
+            WebViewComponentNavigationPolicy.policy(
                 for: URL(string: "https://example.com/promo/index.html")!,
                 isMainFrame: true,
                 expectedOrigin: expectedOrigin
             ),
             .allow
         )
-    }
-
-    // MARK: - HTTP status handling
-
-    func testMainFrameClientAndServerErrorsAreTerminal() {
-        XCTAssertTrue(WebViewNavigationPolicy.isTerminalHTTPError(statusCode: 404, isMainFrame: true))
-        XCTAssertTrue(WebViewNavigationPolicy.isTerminalHTTPError(statusCode: 500, isMainFrame: true))
-    }
-
-    func testMainFrameSuccessAndRedirectStatusesAreNotTerminal() {
-        XCTAssertFalse(WebViewNavigationPolicy.isTerminalHTTPError(statusCode: 200, isMainFrame: true))
-        XCTAssertFalse(WebViewNavigationPolicy.isTerminalHTTPError(statusCode: 304, isMainFrame: true))
-    }
-
-    func testSubFrameErrorsAreNotTerminal() {
-        // A failing sub-resource must not remove the whole component; only main-frame errors do.
-        XCTAssertFalse(WebViewNavigationPolicy.isTerminalHTTPError(statusCode: 404, isMainFrame: false))
-        XCTAssertFalse(WebViewNavigationPolicy.isTerminalHTTPError(statusCode: 500, isMainFrame: false))
     }
 
 }
